move 'stats' classes out of utils
