+++ react-api-app/src/utils.ts
@@ -120,61 +120,3 @@ export const useLocation = () => {
 
   return location;
 };
-
-export class CurrentWeatherStats {
-  temp: number;
-  clouds: number;
-  wind: number;
-  humidity: number;
-  icon: string;
-  message: string;
-
-  constructor(values: Values) {
-    const { cloudCover: clouds, rainIntensity } = values;
-    if (rainIntensity > 0) {
-      this.icon = "/rainy-day.png";
-      if (rainIntensity < 10) {
-        this.message = "Light Rain";
-      } else if (rainIntensity > 30) {
-        this.message = "Moderate Rain";
-      } else {
-        this.message = "Heavy Rain";
-      }
-    } else if (clouds < 20) {
-      this.icon = "/sunny.png";
-      this.message = "Mostly Sunny";
-    } else {
-      this.icon = "/cloudy.png";
-      this.message = "Cloudy";
-    }
-    this.temp = values.temperature;
-    this.clouds = clouds;
-    this.wind = values.windSpeed;
-    this.humidity = values.humidity;
-  }
-}
-
-export class DailyWeatherStats {
-  formattedDate: string;
-  temp: string;
-  icon: string;
-
-  constructor({time, values}: Timeline) {
-    const date = new Date(time);
-    const month = date.toLocaleString('default', { month: 'long' });
-    const numDay = date.getDate(); // 21, for example
-    const weekday = date.toLocaleString('default', { weekday: 'short' });
-    this.formattedDate = `${numDay} ${month}, ${weekday}`
-
-    let high = values.temperatureApparentMax
-    let low = values.temperatureApparentMin
-    this.temp = `${Math.floor(high)}°/${Math.floor(low)}°`
-    if (values.rainAccumulationSum > 0) {
-      this.icon = '/rainy-day.png'
-    } else if (values.cloudCoverAvg < 20) {
-      this.icon = '/sunny.png'
-    } else {
-      this.icon = '/cloudy.png'
-    }
-  }
-}
